Move CoroutineTestRule to test_util module
