+++ app/src/androidTest/java/com/phicdy/mycuration/domain/rss/RssParserTest.kt
@@ -1,7 +1,7 @@
 package com.phicdy.mycuration.domain.rss
 
 import androidx.test.core.app.ApplicationProvider
-import com.phicdy.mycuration.CoroutineTestRule
+import com.phicdy.test.util.CoroutineTestRule
 import com.phicdy.mycuration.data.repository.ArticleRepository
 import com.phicdy.mycuration.data.repository.FilterRepository
 import com.phicdy.mycuration.data.repository.RssRepository
