+++ app/src/androidTest/java/com/phicdy/mycuration/domain/task/GetFeedIconTaskTest.kt
@@ -2,7 +2,7 @@ package com.phicdy.mycuration.domain.task
 
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.phicdy.mycuration.CoroutineTestRule
+import com.phicdy.test.util.CoroutineTestRule
 import com.phicdy.mycuration.data.repository.ArticleRepository
 import com.phicdy.mycuration.data.repository.FilterRepository
 import com.phicdy.mycuration.data.repository.RssRepository
