+++ build.gradle
@@ -103,7 +103,7 @@ subprojects {
             implementation 'androidx.activity:activity-compose:1.8.2'
             implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.6.2'
 
-            testImplementation 'junit:junit:4.13.2'
+            testImplementation libs.junit
             testImplementation 'org.mockito.kotlin:mockito-kotlin:5.1.0'
             testImplementation libs.assertj
         }
