+++ test_util/build.gradle
@@ -13,4 +13,7 @@ android {
 }
 
 dependencies {
+    implementation libs.coroutines.test
+    implementation libs.junit
+    implementation project(':core')
 }
